+++ listapp_webapi-database/leave_group_list.php
@@ -8,14 +8,23 @@ if (isset($_POST['userId']) && isset($_POST['listId'])) {
     $listId = $_POST['listId'];
 
     try {
-        $query = "DELETE FROM listgroup WHERE userid = ? AND listgrouplinkid IN (SELECT id FROM listgrouplink WHERE listid = ?)";
+        $query = "SELECT * FROM users WHERE userid = ? AND archive = 0";
         $statement = $conn->prepare($query);
-        $statement->execute([$userId, $listId]);
+        $statement->execute([$userId]);
+        $user = $statement->fetch(PDO::FETCH_ASSOC);
 
-        if ($statement->rowCount() > 0) {
-            echo "success";
+        if ($user) {
+            $query = "DELETE FROM listgroup WHERE userid = ? AND listgrouplinkid IN (SELECT id FROM listgrouplink WHERE listid = ?)";
+            $statement = $conn->prepare($query);
+            $statement->execute([$userId, $listId]);
+
+            if ($statement->rowCount() > 0) {
+                echo "success";
+            } else {
+                echo "ERROR: Failed to leave the group list.";
+            }
         } else {
-            echo "ERROR: Failed to leave the group list.";
+            echo "ERROR: Invalid user.";
         }
     } catch (PDOException $e) {
         echo "ERROR: Failed to execute the query.";
